requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: Additional useful packages
python-multipart==0.0.6  # For file uploads
email-validator==2.1.0   # For email validation
passlib[bcrypt]==1.7.4   # For password hashing
python-jose[cryptography]==3.3.0  # For JWT tokens
PyJWT==2.8.0  # Alternative JWT library
